Factored SSL context building code into a utility class
